Move component status tracking into `_component_status` sub package
There's going to be status trackers for more component types. Those
can now go into the `_component_status` package.
