+++ src/frequenz/sdk/actor/power_distributing/_component_managers/_battery_manager.py
@@ -19,9 +19,8 @@ from ...._internal._math import is_close_to_zero
 from ....microgrid import connection_manager
 from ....microgrid.component import BatteryData, ComponentCategory, InverterData
 from ....timeseries._quantities import Power
-from .._battery_status_tracker import BatteryStatusTracker
 from .._component_pool_status_tracker import ComponentPoolStatusTracker
-from .._component_status import ComponentPoolStatus
+from .._component_status import BatteryStatusTracker, ComponentPoolStatus
 from .._distribution_algorithm import (
     AggregatedBatteryData,
     BatteryDistributionAlgorithm,
