+++ src/frequenz/sdk/actor/power_distributing/_component_status/_component_status.py
